+++ src/Models/Traits/SignupEmailTrait.php
@@ -8,14 +8,14 @@ use Yii;
 
 trait SignupEmailTrait
 {
-    public function sendSignupEmail(): void
+    public function sendSignupEmail(): bool
     {
         $mail = Yii::$app->getMailer()->compose('@skeleton/../resources/mail/account/create', [
             'user' => $this->user,
             'url' => $this->user->createEmailConfirmationUrl(),
         ]);
 
-        $mail->setSubject(Yii::t('skeleton', 'SIGNUP_EMAIL_SIGN_UP_CONFIRMATION'))
+        return $mail->setSubject(Yii::t('skeleton', 'SIGNUP_EMAIL_SIGN_UP_CONFIRMATION'))
             ->setFrom(Yii::$app->params['email'])
             ->setTo($this->user->email)
             ->send();