+++ src/Test/TestMailer.php
@@ -7,6 +7,11 @@ namespace Hirtz\Skeleton\Test;
 use Hirtz\Skeleton\Mail\Mailer;
 use Hirtz\Skeleton\Mail\Message;
 use Override;
+use Symfony\Component\Mailer\Envelope;
+use Symfony\Component\Mailer\Exception\TransportException;
+use Symfony\Component\Mailer\SentMessage;
+use Symfony\Component\Mailer\Transport\TransportInterface;
+use Symfony\Component\Mime\RawMessage;
 
 class TestMailer extends Mailer
 {
@@ -14,6 +19,11 @@ class TestMailer extends Mailer
      * @var Message[]
      */
     private array $messages = [];
+
+    /**
+     * @var bool whether the transport refuses every message, which `Mail\Mailer` logs and answers with `false`
+     */
+    public bool $isFailing = false;
 
     public function init(): void
     {
@@ -25,8 +35,24 @@ class TestMailer extends Mailer
      * @param Message $message
      */
     #[Override]
-    protected function sendMessage($message): true
+    protected function sendMessage($message): bool
     {
+        if ($this->isFailing) {
+            $this->setTransport(new class () implements TransportInterface {
+                public function send(RawMessage $message, ?Envelope $envelope = null): ?SentMessage
+                {
+                    throw new TransportException('The test transport refuses every message.');
+                }
+
+                public function __toString(): string
+                {
+                    return 'failing://';
+                }
+            });
+
+            return parent::sendMessage($message);
+        }
+
         $this->messages[$this->generateMessageFileName()] = $message;
         return true;
     }
@@ -62,5 +88,6 @@ class TestMailer extends Mailer
     public function reset(): void
     {
         $this->messages = [];
+        $this->isFailing = false;
     }
 }